+++ test/e2e/framework/debug/log_size_monitoring.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package framework
+package debug
 
 import (
 	"bytes"
@@ -27,7 +27,7 @@ import (
 
 	clientset "k8s.io/client-go/kubernetes"
 
-	// TODO: Remove the following imports (ref: https://github.com/kubernetes/kubernetes/issues/81245)
+	"k8s.io/kubernetes/test/e2e/framework"
 	e2essh "k8s.io/kubernetes/test/e2e/framework/ssh"
 )
 
@@ -109,7 +109,7 @@ func (s *LogsSizeDataSummary) PrintHumanReadable() string {
 
 // PrintJSON returns the summary of log size data with JSON format.
 func (s *LogsSizeDataSummary) PrintJSON() string {
-	return PrettyPrintJSON(*s)
+	return framework.PrettyPrintJSON(*s)
 }
 
 // SummaryKind returns the summary of log size data summary.
@@ -158,8 +158,8 @@ func (d *LogsSizeData) addNewData(ip, path string, timestamp time.Time, size int
 // NewLogsVerifier creates a new LogsSizeVerifier which will stop when stopChannel is closed
 func NewLogsVerifier(c clientset.Interface, stopChannel chan bool) *LogsSizeVerifier {
 	nodeAddresses, err := e2essh.NodeSSHHosts(c)
-	ExpectNoError(err)
-	instanceAddress := APIAddress() + ":22"
+	framework.ExpectNoError(err)
+	instanceAddress := framework.APIAddress() + ":22"
 
 	workChannel := make(chan WorkItem, len(nodeAddresses)+1)
 	workers := make([]*LogSizeGatherer, workersNo)
@@ -256,13 +256,13 @@ func (g *LogSizeGatherer) Work() bool {
 	sshResult, err := e2essh.SSH(
 		fmt.Sprintf("ls -l %v | awk '{print $9, $5}' | tr '\n' ' '", strings.Join(workItem.paths, " ")),
 		workItem.ip,
-		TestContext.Provider,
+		framework.TestContext.Provider,
 	)
 	if err != nil {
-		Logf("Error while trying to SSH to %v, skipping probe. Error: %v", workItem.ip, err)
+		framework.Logf("Error while trying to SSH to %v, skipping probe. Error: %v", workItem.ip, err)
 		// In case of repeated error give up.
 		if workItem.backoffMultiplier >= 128 {
-			Logf("Failed to ssh to a node %v multiple times in a row. Giving up.", workItem.ip)
+			framework.Logf("Failed to ssh to a node %v multiple times in a row. Giving up.", workItem.ip)
 			g.wg.Done()
 			return false
 		}
@@ -278,7 +278,7 @@ func (g *LogSizeGatherer) Work() bool {
 		path := results[i]
 		size, err := strconv.Atoi(results[i+1])
 		if err != nil {
-			Logf("Error during conversion to int: %v, skipping data. Error: %v", results[i+1], err)
+			framework.Logf("Error during conversion to int: %v, skipping data. Error: %v", results[i+1], err)
 			continue
 		}
 		g.data.addNewData(workItem.ip, path, now, size)
